pyproject.toml
[tool.poetry]
name = "comeet"
version = "0.1.0"
description = ""
authors = ["VinylStage <mcpe9869@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "^3.6.0"
django = "^4.2.1"
django-extensions = "^3.2.1"
pycodestyle = "^2.10.0"
sqlparse = "^0.4.4"
tzdata = "^2023.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-nextjs = "^2.2.2"
django-cors = "^0.1"
django-cors-headers = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
